design_smell_pipeline/requirements.txt
# Design Smell Detection & Refactoring Pipeline Dependencies
# ===========================================================

# LLM APIs
openai>=1.12.0
google-generativeai>=0.4.0

# GitHub API
PyGithub>=2.1.0

# Git operations
GitPython>=3.1.40

# Configuration
PyYAML>=6.0.1
python-dotenv>=1.0.0

# Java AST parsing
javalang>=0.13.0

# HTTP requests
requests>=2.31.0

# CLI and utilities
click>=8.1.7
rich>=13.7.0
tqdm>=4.66.0

# CSV processing
pandas>=2.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Async support
aiohttp>=3.9.0
asyncio>=3.4.3
